Add files via upload
The problem of the crisis of rare unavailable medicine from abroad can be solved using a simple shipping system to buy medicines from abroad. So, we are thinking about making an application system in which we will make the rare and expensive medicines available anytime from all over the world. As we want to solve this problem, we need to create a system which will save our time & money and provide medicine as soon as possible. The goal of our application is to make a way that can provide rare medicine available anytime. For the second problem of making an appointment the concept of our application which will bring the change is one can make the appointment with any doctors who is appointed in any medical contra’s in all over the country. So, the theme is we are getting all the doctors list in just one application. By this the
appointment system will be much more efficient to use.